refactor(domainFixture): 도메인 fixture를 domain package에 넣어라
