+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.params.provider.NullAndEmptySource;
 
 import java.util.List;
 
-import static kitchenpos.fixture.MenuGroupFixture.menuGroup;
+import static kitchenpos.fixture.domain.MenuGroupFixture.menuGroup;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.Assertions.assertAll;
